Splash Removed, Tools Game Options Started,
Removed the Splash Form; Tried to get it to load all of the configs, but
realized it was not needed due to the fact that they load quickly
anyway.
Tools>Game Options, Started; It does not function as intended as of
right now, but shows you the default settings currently.